Implemented button 'Refresh News'; now it populates the ListView in NewsActivity using a list of SellOffers from the backend.
Reorganized some files: now all Activity classes are in an activities package, and classes that help build GUI objects are in guiHelpers.

Added a layout file for list items in the NewsActivity ListView.

Changed SellOffer on frontend to SellOfferFront; we can use this until both front and backend are using the same code.

UnitsWt is now deprecated.