Added a Detail Activity, called ViewSellOfferActivity, that pops up when you click an item in the NewsActivity ListView.
Changed SellOfferFront to be more in line with what we need for the SellOffer class. Also made the SellOfferFront class implement Parcelable, so that SellOfferFront objects can be passed between activities easily.

Commodity Type is now a String.

Units Weight has been removed from both SellOfferFront and RequestFilteredSellOffers; now we're just going to use $/lb as the standard for everything, and use the frontend to make unit conversions.